2013.4.26
1. Generation of depth
a) In folder 'czt_test', I added 'StereoCamera' and 'StereoVision';
b) There are 2 methods to show recified images and depth image: the 1st one is to use 'VideoSource' which exists originally in DeSTIN, meanwhile I added the new 'getOutput_c1()' to get the 'originalFrame' (maybe I should try to get 'flippedFrame'?); the 2nd one is to use 'StereoCamera';
c) Whatever, the codes could generate depth somehow; but there is an important 'bug'!!! The 'StereoVision' could only be set to (640, 480). If set to other values like 512*512, the cameras will only get 352*288 pixels. It's really frustrating. And still not solved!!!