Bug fix in BeliefExporter. Was crashing because it was trying to delete an uninitialized pointer on startup.
Don't need createMatFile(), instead writeBeliefToMat creates the file
if it doesn't exists.

It leaves the file open, because I believe reopening it each time would
slow things down. Instead we can call closeBeliefFile() after, or
let the destructor do it for us.

Also, it adds the image label / class to the first column of the output
file, so it is easier to correspond the belief to the class label,
plus it works better this way when using the Waffles machine learning tool.

Also moved the file name as an argument to the constructor.